refactor: 공통 예외로 수정
- 여행 -> 엔티티 예외로 수정
