+++ src/main/java/kr/co/fastcampus/travel/common/exception/BaseException.java
@@ -1,5 +1,6 @@
 package kr.co.fastcampus.travel.common.exception;
 
+import kr.co.fastcampus.travel.common.response.ErrorCode;
 import lombok.Getter;
 
 @Getter
