+++ src/main/java/kr/co/fastcampus/travel/common/exception/EntityNotFoundException.java
@@ -1,0 +1,14 @@
+package kr.co.fastcampus.travel.common.exception;
+
+import kr.co.fastcampus.travel.common.response.ErrorCode;
+
+public class EntityNotFoundException extends BaseException {
+
+    public EntityNotFoundException() {
+        super(ErrorCode.COMMON_ENTITY_NOT_FOUND);
+    }
+
+    public EntityNotFoundException(String message) {
+        super(message, ErrorCode.COMMON_ENTITY_NOT_FOUND);
+    }
+}
