--- src/main/java/kr/co/fastcampus/travel/common/exception/TravelDoesNotExistException.java
@@ -1,8 +1,0 @@
-package kr.co.fastcampus.travel.common.exception;
-
-public class TravelDoesNotExistException extends BaseException {
-
-    public TravelDoesNotExistException() {
-        super(ErrorCode.TRAVEL_DOES_NOT_EXIST);
-    }
-}
