+++ src/test/java/kr/co/fastcampus/travel/service/TripServiceTest.java
@@ -7,7 +7,7 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import java.util.List;
 import java.util.stream.IntStream;
-import kr.co.fastcampus.travel.common.exception.TravelDoesNotExistException;
+import kr.co.fastcampus.travel.common.exception.EntityNotFoundException;
 import kr.co.fastcampus.travel.domain.Trip;
 import kr.co.fastcampus.travel.mock.FakeItineraryRepository;
 import kr.co.fastcampus.travel.mock.FakeTripRepository;
@@ -63,7 +63,7 @@ class TripServiceTest {
         // when
         // then
         assertThrows(
-                TravelDoesNotExistException.class,
+                EntityNotFoundException.class,
                 () -> tripService.findTrip(0L)
         );
     }
